# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trading platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserProfileTradingMode = typeof UserProfileTradingMode[keyof typeof UserProfileTradingMode];


export const UserProfileTradingMode = {
  real: 'real',
  demo: 'demo',
} as const;

export interface UserProfile {
  id: number;
  clerkId: string;
  email: string;
  /** @nullable */
  displayName?: string | null;
  tradingMode: UserProfileTradingMode;
  createdAt: string;
}

export type UserProfileUpdateTradingMode = typeof UserProfileUpdateTradingMode[keyof typeof UserProfileUpdateTradingMode];


export const UserProfileUpdateTradingMode = {
  real: 'real',
  demo: 'demo',
} as const;

export interface UserProfileUpdate {
  displayName?: string;
  tradingMode?: UserProfileUpdateTradingMode;
}

export interface WalletBalance {
  realBalance: number;
  demoBalance: number;
  currency: string;
}

export type DepositInputMethod = typeof DepositInputMethod[keyof typeof DepositInputMethod];


export const DepositInputMethod = {
  card: 'card',
  bank_transfer: 'bank_transfer',
  crypto: 'crypto',
} as const;

export interface DepositInput {
  amount: number;
  method: DepositInputMethod;
}

export type WithdrawInputMethod = typeof WithdrawInputMethod[keyof typeof WithdrawInputMethod];


export const WithdrawInputMethod = {
  bank_transfer: 'bank_transfer',
  crypto: 'crypto',
} as const;

export interface WithdrawInput {
  amount: number;
  method: WithdrawInputMethod;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  deposit: 'deposit',
  withdrawal: 'withdrawal',
  trade_buy: 'trade_buy',
  trade_sell: 'trade_sell',
  fee: 'fee',
} as const;

export type TransactionStatus = typeof TransactionStatus[keyof typeof TransactionStatus];


export const TransactionStatus = {
  pending: 'pending',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Transaction {
  id: number;
  type: TransactionType;
  amount: number;
  currency: string;
  status: TransactionStatus;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export type PortfolioMode = typeof PortfolioMode[keyof typeof PortfolioMode];


export const PortfolioMode = {
  real: 'real',
  demo: 'demo',
} as const;

export type PositionSide = typeof PositionSide[keyof typeof PositionSide];


export const PositionSide = {
  buy: 'buy',
  sell: 'sell',
} as const;

export type PositionMode = typeof PositionMode[keyof typeof PositionMode];


export const PositionMode = {
  real: 'real',
  demo: 'demo',
} as const;

export interface Position {
  id: number;
  symbol: string;
  assetName: string;
  side: PositionSide;
  quantity: number;
  entryPrice: number;
  currentPrice: number;
  pnl: number;
  pnlPercent: number;
  mode: PositionMode;
  openedAt: string;
}

export interface Portfolio {
  totalValue: number;
  totalPnl: number;
  totalPnlPercent: number;
  positions: Position[];
  mode: PortfolioMode;
}

export type PortfolioSummaryMode = typeof PortfolioSummaryMode[keyof typeof PortfolioSummaryMode];


export const PortfolioSummaryMode = {
  real: 'real',
  demo: 'demo',
} as const;

export interface PortfolioSummary {
  totalValue: number;
  dayChange: number;
  dayChangePercent: number;
  totalPnl: number;
  totalPnlPercent: number;
  winRate: number;
  totalTrades: number;
  openPositions: number;
  /** @nullable */
  bestTrade?: number | null;
  /** @nullable */
  worstTrade?: number | null;
  mode: PortfolioSummaryMode;
}

export type TradeSide = typeof TradeSide[keyof typeof TradeSide];


export const TradeSide = {
  buy: 'buy',
  sell: 'sell',
} as const;

export type TradeOrderType = typeof TradeOrderType[keyof typeof TradeOrderType];


export const TradeOrderType = {
  market: 'market',
  limit: 'limit',
  stop_loss: 'stop_loss',
  take_profit: 'take_profit',
} as const;

export type TradeStatus = typeof TradeStatus[keyof typeof TradeStatus];


export const TradeStatus = {
  pending: 'pending',
  open: 'open',
  closed: 'closed',
  cancelled: 'cancelled',
} as const;

export type TradeMode = typeof TradeMode[keyof typeof TradeMode];


export const TradeMode = {
  real: 'real',
  demo: 'demo',
} as const;

export interface Trade {
  id: number;
  symbol: string;
  assetName: string;
  side: TradeSide;
  orderType: TradeOrderType;
  quantity: number;
  price: number;
  /** @nullable */
  executedPrice?: number | null;
  /** @nullable */
  stopLoss?: number | null;
  /** @nullable */
  takeProfit?: number | null;
  status: TradeStatus;
  /** @nullable */
  pnl?: number | null;
  mode: TradeMode;
  createdAt: string;
  /** @nullable */
  closedAt?: string | null;
}

export type TradeInputSide = typeof TradeInputSide[keyof typeof TradeInputSide];


export const TradeInputSide = {
  buy: 'buy',
  sell: 'sell',
} as const;

export type TradeInputOrderType = typeof TradeInputOrderType[keyof typeof TradeInputOrderType];


export const TradeInputOrderType = {
  market: 'market',
  limit: 'limit',
  stop_loss: 'stop_loss',
  take_profit: 'take_profit',
} as const;

export type TradeInputMode = typeof TradeInputMode[keyof typeof TradeInputMode];


export const TradeInputMode = {
  real: 'real',
  demo: 'demo',
} as const;

export interface TradeInput {
  symbol: string;
  side: TradeInputSide;
  orderType: TradeInputOrderType;
  quantity: number;
  price: number;
  /** @nullable */
  stopLoss?: number | null;
  /** @nullable */
  takeProfit?: number | null;
  mode: TradeInputMode;
}

export type AssetCategory = typeof AssetCategory[keyof typeof AssetCategory];


export const AssetCategory = {
  crypto: 'crypto',
  forex: 'forex',
  stocks: 'stocks',
  commodities: 'commodities',
} as const;

export interface Asset {
  symbol: string;
  name: string;
  category: AssetCategory;
  currentPrice: number;
  change24h: number;
  change24hPercent: number;
  volume24h: number;
  /** @nullable */
  logoUrl?: string | null;
}

export interface AssetPrice {
  symbol: string;
  price: number;
  bid: number;
  ask: number;
  change24h: number;
  change24hPercent: number;
  high24h: number;
  low24h: number;
  volume24h: number;
  updatedAt: string;
}

export interface Candle {
  time: number;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
}

export interface WatchlistItem {
  symbol: string;
  name: string;
  category: string;
  currentPrice: number;
  change24h: number;
  change24hPercent: number;
}

export interface WatchlistInput {
  symbol: string;
}