# lib/api-zod/src/generated/types/position.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trading platform API
 * OpenAPI spec version: 0.1.0
 */
import type { PositionMode } from './positionMode';
import type { PositionSide } from './positionSide';

export interface Position {
  id: number;
  symbol: string;
  assetName: string;
  side: PositionSide;
  quantity: number;
  entryPrice: number;
  currentPrice: number;
  pnl: number;
  pnlPercent: number;
  mode: PositionMode;
  openedAt: string;
}